backend/telegram-bot/requirements.txt
pydantic==2.5.0
psycopg2-binary==2.9.9
openai==1.54.0
httpx==0.27.0
requests==2.31.0